Move methods out of concern
