+++ app/models/invoice.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class Invoice < ApplicationRecord
-  include ::InvoiceConcern
+  after_update :lock_discounted_prices
 
   belongs_to :customer
   has_many :invoice_items, dependent: :destroy
@@ -37,5 +37,36 @@ class Invoice < ApplicationRecord
 
   def discounts_and_discounted_total
     discounted_info
+  end
+
+  def discount_and_total
+    inv_item_discounts.group_by(&:id).map do |inv_item_id, discount|
+      max_d = discount.max_by(&:percentage)
+      amount_off = max_d.revenue * max_d.percentage.fdiv(100)
+
+      discount_info = {
+        id: max_d.discount,
+        amount_off: amount_off
+      }
+
+      [inv_item_id, discount_info]
+    end.to_h
+  end
+
+  def discounted_info
+    result = discount_and_total
+    price_off = result.sum do |_inv, dis|
+      dis[:amount_off]
+    end
+    result.merge({ discounted_total: total_revenue - price_off })
+  end
+
+  def lock_discounted_prices
+    return unless completed?
+    Invoice.transaction do
+      invoice_items.each do |inv_item|
+        inv_item.lock_discounted_price
+      end
+    end
   end
 end
